fix(repo): remove 2-space prefix in repo JSON
I'm planning on removing this code in favor of https://github.com/armory-io/plugin-metadata-updater. 

While both tools (i.e., this code and the GitHub action) are in use, it would be best if they used the same JSON format. This code includes a two space "prefix" for each line, which was probably never intended.